# api/v1/openapi.yaml
openapi: 3.0.3
info:
  title: Go Tasks API
  description: |
    Task and habit management API. Users register and manage their own tasks,
    categories, daily logs, and answers. All data is scoped to the authenticated user.

    ## Authentication

    Protected endpoints require a valid JWT access token in the `Authorization` header
    using the Bearer scheme. Access tokens are RS256-signed and valid for 15 minutes.

    **Token Claims:**
    - `sub` — user ID (UUID)
    - `iss` — issuer (configured via JWT_ISSUER)
    - `aud` — audience (configured via JWT_AUDIENCE)
    - `exp` — expiration time
    - `nbf` — not before time
    - `iat` — issued at time
    - `jti` — unique token ID (UUID)

    ## Refresh Tokens

    Refresh tokens are valid for 1 hour and support rotation. On use, the old token
    is deleted and a new one is issued. Logout revokes the refresh token and adds
    the access token's jti to a blocklist.

  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Development server
security:
  - BearerAuth: []
...
paths:
...
  /api/v1/occurrences:
    get:
      tags:
        - Occurrence
      summary: List occurrences
      description: |
        Returns task occurrences for a given date or date range. Occurrences are
        generated on demand based on task schedules.

        **Required:** Either `date` OR both `start_date` and `end_date` must be provided.
        Returns 400 if no date parameters are supplied.

        Suppressed occurrences are included in the response with `is_suppressed: true`
        so the UI can display them as skipped.
      operationId: listOccurrences
      parameters:
        - name: date
          in: query
          description: Single day query (YYYY-MM-DD). Required if start_date/end_date not provided.
          schema:
            type: string
            format: date
        - name: start_date
          in: query
          description: Range query start (YYYY-MM-DD). Required with end_date if date not provided.
          schema:
            type: string
            format: date
        - name: end_date
          in: query
          description: Range query end (YYYY-MM-DD). Required with start_date if date not provided.
          schema:
            type: string
            format: date
      responses:
        "200":
          description: List of occurrences with details
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OccurrenceWithDetails"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/occurrences/{id}/suppress:
    post:
      tags:
        - Occurrence
      summary: Suppress an occurrence
      description: Marks this occurrence as suppressed (skipped) for this day only
      operationId: suppressOccurrence
      parameters:
        - $ref: "#/components/parameters/ID"
      responses:
        "204":
          description: Occurrence suppressed
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: Occurrence already suppressed
  /api/v1/occurrences/{id}/unsuppress:
    post:
      tags:
        - Occurrence
      summary: Unsuppress an occurrence
      description: Removes suppression from an occurrence, making it active again
      operationId: unsuppressOccurrence
      parameters:
        - $ref: "#/components/parameters/ID"
      responses:
        "204":
          description: Occurrence unsuppressed
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: Occurrence is not suppressed
  /api/v1/occurrences/{id}/answer:
    post:
      tags:
        - Occurrence
      summary: Submit an answer
      description: Submit or update the answer for an occurrence
      operationId: submitAnswer
      parameters:
        - $ref: "#/components/parameters/ID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnswerRequest"
      responses:
        "200":
          description: Answer saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskAnswer"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"

  # Daily log endpoints
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: RS256-signed JWT access token
  parameters:
    ID:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  schemas:
...
    Task:
      type: object
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        category_id:
          type: string
          format: uuid
        name:
          type: string
          maxLength: 200
        description:
          type: string
          maxLength: 1000
        answer_type:
          type: string
          enum: [string, integer, boolean, select]
        is_active:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    TaskSelectOption:
      type: object
      properties:
        id:
          type: string
          format: uuid
        task_id:
          type: string
          format: uuid
        value:
          type: string
          maxLength: 100
        position:
          type: integer
        created_at:
          type: string
          format: date-time
...
    TaskOccurrence:
      type: object
      properties:
        id:
          type: string
          format: uuid
        task_id:
          type: string
          format: uuid
        schedule_id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        occurrence_date:
          type: string
          format: date
        scheduled_time:
          type: string
          format: time
        is_suppressed:
          type: boolean
        created_at:
          type: string
          format: date-time
    TaskAnswer:
      type: object
      properties:
        id:
          type: string
          format: uuid
        occurrence_id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        answer_string:
          type: string
          maxLength: 500
        answer_integer:
          type: integer
        answer_boolean:
          type: boolean
        answer_select:
          type: string
          format: uuid
        answered_at:
          type: string
          format: date-time
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    OccurrenceWithDetails:
      type: object
      properties:
        occurrence:
          $ref: "#/components/schemas/TaskOccurrence"
        task:
          $ref: "#/components/schemas/Task"
        select_options:
          type: array
          items:
            $ref: "#/components/schemas/TaskSelectOption"
        answer:
          $ref: "#/components/schemas/TaskAnswer"
    AnswerRequest:
      type: object
      properties:
        answer_string:
          type: string
          maxLength: 500
        answer_integer:
          type: integer
        answer_boolean:
          type: boolean
        answer_select:
          type: string
          format: uuid

    # Daily log schemas
...
  responses:
    BadRequest:
      description: Bad request
    Unauthorized:
      description: Unauthorized
    NotFound:
      description: Resource not found